+++ apps/docs/app/ui/layout/sidebar/Sidebar.tsx
@@ -1,11 +1,13 @@
 "use client";
 
-import Overlay from "@/components/overlay/Overlay";
-import getPageLinks from "@/app/lib/getPageLinks";
 import clsx from "clsx";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { useEffect, useRef } from "react";
+import { useSidebar } from "@/context/sidebar/SidebarProvider";
+
+import Overlay from "@/components/overlay/Overlay";
+import getPageLinks from "@/app/lib/getPageLinks";
 
 import type { Data } from "@/app/lib/getPageLinks";
 
@@ -13,59 +15,50 @@ import "./sidebar.css";
 
 interface SidebarProps {
     data: Data[];
-    isOpen: boolean;
-    onClose: () => void;
 }
 
-const Sidebar = ({ data, isOpen, onClose }: SidebarProps) => {
+const Sidebar = ({ data }: SidebarProps) => {
     const sidebarRef = useRef<HTMLDivElement>(null);
     const links = getPageLinks(data, { order: ["getting-started", "semantic", "core", "react-icons", "svg"] });
     const pathName = usePathname();
+    const { toggleSidebar, isSidebarOpen } = useSidebar();
 
     useEffect(() => {
         const handleClickOutside = (event: MouseEvent) => {
             if (sidebarRef.current && !sidebarRef.current.contains(event.target as Node)) {
-                onClose();
+                toggleSidebar();
             }
         };
 
         const handleLockScroll = () => {
-            const shouldLockScroll = window.innerWidth <= 768 && isOpen;
+            const shouldLockScroll = window.innerWidth <= 768 && isSidebarOpen;
             document.body.style.overflow = shouldLockScroll ? "hidden" : "visible";
-        };
-
-        const handleKeyDown = (event: KeyboardEvent) => {
-            if (event.key === "Escape") {
-                onClose();
-            }
         };
 
         const handleWindowResize = () => {
             if (window.innerWidth > 768) {
-                onClose();
+                toggleSidebar();
             } else {
                 handleLockScroll();
             }
         };
 
-        if (isOpen) {
+        if (isSidebarOpen) {
             handleLockScroll();
             document.addEventListener("click", handleClickOutside);
-            document.addEventListener("keydown", handleKeyDown);
             window.addEventListener("resize", handleWindowResize);
         }
 
         return () => {
             document.body.style.overflow = "auto";
             document.removeEventListener("click", handleClickOutside);
-            document.removeEventListener("keydown", handleKeyDown);
             window.removeEventListener("resize", handleWindowResize);
         };
-    }, [isOpen, onClose]);
+    }, [isSidebarOpen, toggleSidebar]);
 
     const handleLinkClick = () => {
         if (window.innerWidth <= 768) {
-            onClose();
+            toggleSidebar();
         }
     };
 
@@ -98,8 +91,11 @@ const Sidebar = ({ data, isOpen, onClose }: SidebarProps) => {
 
     return (
         <>
-            <Overlay isOpen={isOpen}></Overlay>
-            <nav className={clsx("hd-sidebar", isOpen && "hd-sidebar--open")} aria-label="Sidebar" ref={sidebarRef}>
+            <Overlay isOpen={isSidebarOpen}></Overlay>
+            <nav className={clsx("hd-sidebar", isSidebarOpen && "hd-sidebar--open")}
+                aria-label="Sidebar"
+                ref={sidebarRef}
+            >
                 <div className="hd-sidebar__container">
                     {linkItems}
                 </div>