+++ apps/docs/app/ui/layout/subHeader/SubHeader.stories.tsx
@@ -33,8 +33,6 @@ export const Default: Story = {
                 url: "#section-3",
                 id: "section-3"
             }
-        ],
-        toggleOpenState: () => {
-        }
+        ]
     }
 };