+++ packages/components/src/badge/src/BadgeContext.ts
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+import type { ContextValue } from "react-aria-components";
+
+import type { BadgeProps } from "./Badge.tsx";
+
+export const BadgeContext = createContext<ContextValue<BadgeProps, HTMLSpanElement>>({});
+
+BadgeContext.displayName = "BadgeContext";