+++ packages/components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hopper-ui/components
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [a92e653]
+  - @hopper-ui/icons@2.1.1
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [75db68e]
+  - @hopper-ui/icons@2.1.0
+
 ## 1.0.0
 
 ### Major Changes